refactor: добавил метод addEvent вместо трех для фиксации события, правки по codeStyle
