+++ src/main/java/ru/yandex/practicum/filmorate/repository/UserRepository.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.Constants;
 import ru.yandex.practicum.filmorate.util.FilmRowMapper;
 
 import java.time.ZoneId;
@@ -22,7 +21,7 @@ public class UserRepository {
 
     private JdbcTemplate template;
     private FilmRowMapper filmRowMapper;
-    private FeedRepository feedRepository;
+    private EventRepository eventRepository;
 
     public User create(User user) {
         if (user.getName().isEmpty()) {
@@ -84,7 +83,7 @@ public class UserRepository {
                 "insert into follows (following_id, followed_id) values(?, ?)",
                 friendId, userId);
         log.info("subscribe user '{}' to user '{}'", userId, friendId);
-        this.makeRecordOnAdd(userId, friendId);
+        eventRepository.addEvent(userId, friendId, "FRIEND", "ADD");
         return getById(userId);
     }
 
@@ -95,7 +94,7 @@ public class UserRepository {
                 "delete from follows where following_id = ? and followed_id = ?",
                 friendId, userId);
         log.info("unsubscribe user '{}' from user '{}'", userId, friendId);
-        this.makeRecordOnRemove(userId, friendId);
+        eventRepository.addEvent(userId, friendId, "FRIEND", "REMOVE");
         return getById(userId);
     }
 
@@ -200,13 +199,5 @@ public class UserRepository {
     public void delUserById(int userId) {
         template.update("DELETE FROM users WHERE id=?", userId);
         log.info("deleted user by id '{}'", userId);
-    }
-
-    private void makeRecordOnAdd(Integer userId, Integer friendId) {
-        feedRepository.recordAddEvent(userId,Constants.EVENT_TYPE_FRIEND, friendId, Constants.ADD_OPERATION);
-    }
-
-    private void makeRecordOnRemove(Integer userId, Integer friendId) {
-        feedRepository.recordAddEvent(userId,Constants.EVENT_TYPE_FRIEND, friendId, Constants.REMOVE_OPERATION);
     }
 }
