--- src/main/java/ru/yandex/practicum/filmorate/storage/Constants.java
@@ -1,12 +1,0 @@
-package ru.yandex.practicum.filmorate.storage;
-
-public class Constants {
-
-    public static final String REMOVE_OPERATION = "REMOVE";
-    public static final String ADD_OPERATION = "ADD";
-    public static final String UPDATE_OPERATION = "UPDATE";
-
-    public static final String EVENT_TYPE_LIKE = "LIKE";
-    public static final String EVENT_TYPE_REVIEW = "REVIEW";
-    public static final String EVENT_TYPE_FRIEND = "FRIEND";
-}
